xz: Mask control characters (actually only bytes) in untrusted strings
Filenames are the dangerous thing as they could contain weird bytes
and globbing can easy pass such names to xz. This commit calls the
masking function on a few other strings from the command line too.

The lack of this feature has been listed in TODO since 2009
5f6dddc6c911df02ba660564e78e6de80947c947.

It would be more than nice to have a kernel option to ban *creation*
of files with bytes 0x01-0x1F and 0x7F. It wouldn't be a complete
fix but it would be a start. The problem was recognized long ago.
See: https://dwheeler.com/essays/fixing-unix-linux-filenames.html